+++ CHANGELOG.md
@@ -9,6 +9,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Changed
 
+## [v1.0.0-beta.3] - 2023-06-19
+### Changed
+- Fix of proc-macro attributes deletion #5
+
 ## [v1.0.0-beta.2] - 2023-04-02
 ### Added
 - Support of `poem-openapi` #3